Fixed issue #3 and finished restoration
The xmls changes are done. All library symbols are now organized and have a proper name. Same for layers. Every layer has a single shape/object. The layer name will always be the stage instance name of whatever is on the layer. If the object does not have an instance name, then a one/word descriptive name is given. If its so basic or obvious what the layer content is, I just named it by the Library symbol type, like `Shape` or `TextField`. I also fixed issue 3. Ill commit/PR/close in a soon.